game logic and some features are complete
Today's changes (6/10/19):
Fixed not colliding ground bug by unchecking isTrigger in ground collider and adding a polygon collider to charizard. (and checked autoTile in ground collider).
Created charizard's flying animation, and dead charizard.
reScaled charizard (0.7,0.7,1) and the columns(0.15,0.175,1) so that they'll fit better.
Game-Over screen Is ready (besides front end modifications), and a click on the screen is restarting the game (with an animation of text that is fading in and out in 0.25seconds rate and switching colors.
Added some scripts and turned the columns into a prefab with a collidable so that the game logic is now complete and ready for adding more features.