Respawned boosts by boosts pool.
Respawned boosts by boosts pool.
Changed the boosts script so that the boosts will act as they should (hide when being triggered, and won't be active when they're triggered until they're respawned).
The boosts are spawned in different rates (randomized rates) and they are spawned in different x positions so that they'll be between the columns and won't collide them.
changed column prefab so that the columns would be in 0 position in x axes.