fireball and enemy are working
fireball is being created by charizard's mouth, and has speed so it seems like it being shot.
the fireball is being destroyed when it hit a column or an enemy.
there is an enemyPool that is constantly using the former enemies.
created zapdos animation.
when an enemy hits charizard, charizard is dead and when it is getting shot it is dead and the score is updated.
did some fmodifications in boost script, and added shot function to player.